test(core): cover older Angular runtime services
Why:
Effects and after-render callbacks have existed since Angular 16, and resource() since
Angular 19, but their regressions only covered Angular 20+.

What:
Split the runtime tests at their API boundaries, add an Angular 16 EffectManager render
regression, and use cross-version lifecycle flushing for zoned and zoneless profiles.

Where:
Extend the compatibility spread from Angular 16 through 22 without adding speculative
never-mock exceptions.

Related to #14613 and follow-up to #14614 and #14619.
