+++ test-spread.conf
@@ -59,14 +59,16 @@ file|tests/issue-14528/*.ts|versions=6+
 file|tests/issue-14544/test.spec.ts|features=injectFn
 file|tests/issue-14544/service.spec.ts|versions=22
 file|tests/issue-14560/*.ts|features=injectFn
-file|tests/issue-14613/*.ts|versions=20+
+file|tests/issue-14613/*.ts|versions=17+
 file|tests/issue-9397/*.ts|features=injectFn
 file|tests/issue-10760/*.ts|features=injectFn
 file|examples/ngMocksGlobalMock/*.spec.ts|features=injectFn
 file|examples/TestRoutingGuard/test.spec.ts|versions=6+
 file|examples/TestRoutingResolver/test.spec.ts|versions=5+
 file|tests/mock-service/observable.spec.ts|versions=6+
-file|tests/provider-never-mock/angular-20-plus.spec.ts|versions=20+
+file|tests/provider-never-mock/angular-16-plus-after-render.spec.ts|versions=16+
+file|tests/provider-never-mock/angular-16-effect-manager.spec.ts|versions=16
+file|tests/provider-never-mock/angular-19-plus-resource.spec.ts|features=resource
 file|tests/issue-1596/test.spec.ts|versions=5-21
 file|tests/issue-296/test.spec.ts|versions=5-21
 file|tests/issue-736/test.spec.ts|versions=5-21
