+++ tests/issue-14613/test.spec.ts
@@ -15,16 +15,15 @@ class TargetInterceptor {
 }
 
 // @see https://github.com/help-me-mom/ng-mocks/issues/14613
-// Angular root effects require the framework's EffectScheduler. MockBuilder
-// used to replace it with an empty class mock, so the first effect creation
-// failed because the mock scheduler did not implement `add`.
+// Angular 17+ root effects require the framework's EffectScheduler. MockBuilder
+// used to replace it with an empty class mock, so effect registration failed.
 describe('issue-14613', () => {
   beforeEach(() => MockBuilder(TargetInterceptor));
 
   it('creates and runs an effect registered by a root service', () => {
     const target = ngMocks.findInstance(TargetInterceptor);
 
-    TestBed.tick();
+    TestBed.flushEffects();
 
     expect(target.called).toBe(true);
   });
